requirements.txt
# Core dependencies
markdown-it-py>=3.0.0
orjson>=3.9.0
neo4j>=5.15.0
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2

# Optional fallback
faiss-cpu>=1.7.4

# CLI
click>=8.1.7
pyyaml>=6.0.1

# Development
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.12.1
isort>=5.13.2
flake8>=6.1.0
mypy>=1.8.0

# Utilities
python-dateutil>=2.8.2
hashlib-compat>=1.0.1

